Revisit naming and object resource support
Because of high frequency of the "supplier" term, the ResourceSupplier was
renamed to ResourceProvider (clashing a bit with ResourceBundleProvider,
but not so much, hopefully).

ResourceDefinition annotation had somewhat misleading name as well, after
incorporating object resources, therefore it was renamed to
ResourceString, which should fit more its purpose.

The name changes were reflected in the documentation and a few other minor
improvements were made.